Renamed the todo screen to the list screen for more clarity
