+++ todo-list-client/src/components/home/newListForm/NewListForm.component.tsx
@@ -6,18 +6,18 @@ import styles from './NewListForm.module.scss';
 import { FormProvider } from 'react-hook-form';
 
 const NewListForm = () => {
-    const createNewListData = useCreateNewList();
+  const createNewListData = useCreateNewList();
 
-    return (
-        <FormProvider {...createNewListData.formData}>
-            <form className={styles.newListForm} onSubmit={createNewListData.onSubmit}>
-                <div className={styles.formFields}>
-                    <NewListFormInputFields />
-                    <SubmitButton props={{ variant: VariantsEnum.SECONDARY }} />
-                </div>
-            </form>
-        </FormProvider>
-    );
+  return (
+    <FormProvider {...createNewListData.formData}>
+      <form className={styles.newListForm} onSubmit={createNewListData.onSubmit}>
+        <div className={styles.formFields}>
+          <NewListFormInputFields />
+          <SubmitButton props={{ variant: VariantsEnum.SECONDARY }} />
+        </div>
+      </form>
+    </FormProvider>
+  );
 };
 
 export default NewListForm;
