+++ todo-list-client/src/components/home/newListForm/NewListForm.constants.tsx
@@ -1,8 +1,8 @@
 export const NEW_LIST_FORM_INPUT_FIELDS = [
-    {
-        type: 'text',
-        name: 'title',
-        placeholder: 'todo.title',
-        required: true,
-    },
+  {
+    type: 'text',
+    name: 'title',
+    placeholder: 'todo.title',
+    required: true,
+  },
 ];
