+++ todo-list-client/src/components/home/newListForm/NewListForm.module.scss
@@ -2,16 +2,16 @@
 @use '../../../design/shared/radiuses';
 
 .newListForm {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    padding-top: 0;
-    border-radius: radiuses.$md;
-    animation: fade-in 0.8s ease-in-out;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  padding-top: 0;
+  border-radius: radiuses.$md;
+  animation: fade-in 0.8s ease-in-out;
 }
 
 .formFields {
-    display: flex;
-    flex-direction: column;
-    gap: spacing.$sm;
+  display: flex;
+  flex-direction: column;
+  gap: spacing.$sm;
 }
