+++ todo-list-client/src/components/home/newListForm/NewListForm.schema.tsx
@@ -1,5 +1,5 @@
 import * as yup from 'yup';
 
 export const NewListSchema = yup.object({
-    title: yup.string().required('todo.title_required')
+  title: yup.string().required('todo.title_required'),
 });
