+++ todo-list-client/src/components/list/progress/Progress.constants.tsx
@@ -1,0 +1,3 @@
+export const TODO_PROGRESS_PERCENTAGE = {
+  percentage: 67,
+};
