+++ todo-list-client/src/components/list/progress/Progress.module.scss
@@ -1,0 +1,46 @@
+@use '../../../design/shared/spacing';
+@use '../../../design/shared/radiuses';
+@use '../../../design/shared/colors';
+@use '../../../design/shared/shadows';
+@use '../../../design/shared/transitions';
+@use '../../../design/shared/typography';
+
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: spacing.$sm;
+  width: 80%;
+  max-width: 1000px;
+}
+
+.label {
+  color: var(--neutral-two);
+  text-align: center;
+}
+
+.bar {
+  position: relative;
+  height: 24px;
+  background-color: colors.$faded-dark-grey;
+  border-radius: radiuses.$md;
+  overflow: hidden;
+  box-shadow: shadows.$xs;
+
+  .percentage {
+    width: 100%;
+    height: 100%;
+    color: colors.$white;
+    font-size: typography.$xxs;
+    position: absolute;
+    top: 0;
+    text-align: center;
+  }
+}
+
+.fill {
+  background: colors.$success;
+  height: 100%;
+  border-radius: radiuses.$md;
+  width: 0%;
+  @include transitions.transition(width, transitions.$extraSlow);
+}
