+++ todo-list-client/src/components/list/progress/Progress.types.ts
@@ -1,0 +1,7 @@
+export type TProgressProps = {
+  props: {
+    percentage: number;
+  };
+};
+
+export type TUseTodoListPercentageCompletedProps = TProgressProps;
