+++ todo-list-client/src/components/list/progress/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Progress.component';
