+++ todo-list-client/src/components/list/title/Title.constants.tsx
@@ -1,0 +1,4 @@
+export const TODO_TITLE_DATA = {
+  title: 'My Morning Routine',
+  date: '04/22/2025',
+};
