+++ todo-list-client/src/components/list/title/Title.types.ts
@@ -1,0 +1,6 @@
+export type TTitleProps = {
+  props: {
+    title: string;
+    date?: string;
+  };
+};
