+++ todo-list-client/src/components/list/title/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Title.component';
