+++ todo-list-client/src/components/shared/input/Input.module.scss
@@ -48,7 +48,7 @@
     margin-top: spacing.$xs;
     text-align: left;
 
-    [dir="rtl"] & {
+    [dir='rtl'] & {
       text-align: right;
     }
   }
