+++ todo-list-client/src/components/shared/modal/Modal.types.ts
@@ -1,17 +1,17 @@
-import { VariantsEnum } from "../../../enums/variants";
+import { VariantsEnum } from '../../../enums/variants';
 
 export type TModalProps = {
-    props: {
-        isOpen: boolean | undefined;
-        onClose: () => void;
-        title: string;
-        variant?: VariantsEnum;
-    },
-    children: React.ReactNode;
+  props: {
+    isOpen: boolean | undefined;
+    onClose: () => void;
+    title: string;
+    variant?: VariantsEnum;
+  };
+  children: React.ReactNode;
 };
 
 export type TUseToggleModalProps = {
-    props: {
-        isOpen: boolean | undefined;
-    };
+  props: {
+    isOpen: boolean | undefined;
+  };
 };
