+++ todo-list-client/src/components/shared/modal/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Modal.component";
+export { default } from './Modal.component';
