+++ todo-list-client/src/components/shared/submitButton/SubmitButton.module.scss
@@ -1,19 +1,19 @@
-@use "../../../design/components/buttons";
-@use "../../../design/shared/colors";
+@use '../../../design/components/buttons';
+@use '../../../design/shared/colors';
 
 .submitButton {
-    width: 100%;
-    @include buttons.primary;
+  width: 100%;
+  @include buttons.primary;
 
-    &[data-variant="primary"] {
-        background-color: colors.$primary;
-    }
+  &[data-variant='primary'] {
+    background-color: colors.$primary;
+  }
 
-    &[data-variant="secondary"] {
-        background-color: colors.$secondary;
-    }
+  &[data-variant='secondary'] {
+    background-color: colors.$secondary;
+  }
 
-    &[data-variant="success"] {
-        background-color: colors.$success;
-    }
+  &[data-variant='success'] {
+    background-color: colors.$success;
+  }
 }
