--- todo-list-client/src/components/todo/progress/Progress.constants.tsx
@@ -1,3 +1,0 @@
-export const TODO_PROGRESS_PERCENTAGE = {
-    percentage: 67
-};
