--- todo-list-client/src/components/todo/progress/Progress.types.ts
@@ -1,7 +1,0 @@
-export type TProgressProps = {
-    props: {
-        percentage: number;
-    }
-};
-
-export type TUseTodoListPercentageCompletedProps = TProgressProps;
