--- todo-list-client/src/components/todo/progress/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Progress.component";
