--- todo-list-client/src/components/todo/title/Title.constants.tsx
@@ -1,4 +1,0 @@
-export const TODO_TITLE_DATA = { 
-    title: "My Morning Routine", 
-    date: "04/22/2025" 
-};
