--- todo-list-client/src/components/todo/title/Title.module.scss
@@ -1,22 +1,0 @@
-@use "../../../design/shared/animations";
-@use "../../../design/shared/spacing";
-@use "../../../design/shared/colors";
-@use "../../../design/shared/typography";
-
-.wrapper {
-    width: 80%;
-    max-width: 1000px;
-    margin-top: spacing.$lg;
-    @include animations.fade-in;
-
-    .title {
-        font-size: typography.$lg;
-        color: var(--neutral-two);
-        font-weight: bold;
-    }
-    
-    .date {
-        font-size: typography.$sm;
-        color: colors.$dark-grey;
-    }
-}
