--- todo-list-client/src/components/todo/title/Title.types.ts
@@ -1,6 +1,0 @@
-export type TTitleProps = {
-    props: {
-        title: string;
-        date?: string;
-    }
-};
