--- todo-list-client/src/components/todo/title/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Title.component';
