+++ todo-list-client/src/enums/variants.ts
@@ -1,6 +1,6 @@
 export enum VariantsEnum {
-    PRIMARY = "primary",
-    SECONDARY = "secondary",
-    SUCCESS = "success",
-    ERROR = "error"
+  PRIMARY = 'primary',
+  SECONDARY = 'secondary',
+  SUCCESS = 'success',
+  ERROR = 'error',
 }
